correção da funcionalidade de salvar o usuário e editora do livro; Agora envia os IDs das entidades que são recuperadas no controller e anexadas a entidade do Livro;

+++ src/pb/service/EditoraService.java
@@ -7,7 +7,7 @@ import javax.persistence.PersistenceContext;
 import pb.entity.Editora;
 
 @Stateless
-public class EditoraService extends AbstractPersistence<Editora, Long> {
+public class EditoraService extends AbstractPersistence<Editora, Integer> {
 
 	/**
 	 * O container injeta a referência para o EntityManager.
